requirements.txt
streamlit
pandas
google-api-python-client
oauth2client
requests
langchain
